requirements.txt
asgiref==3.9.1
azure-core==1.35.1
azure-identity==1.25.1
azure-storage-blob==12.26.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.5
django-cors-headers==4.8.0
django-storages==1.14.6
djangorestframework==3.16.1
idna==3.10
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.11.0
